✨ Plotly 기본틀 생성 및 result API 분리
plotly 기본틀 생성 및 result API 분리
