+++ app/templates/results.html
@@ -41,28 +41,9 @@
                 <h3 class="text-center">연령별 사용자와 동일한 답변 선택자 비율</h3>
                 <div id="age_distribution_chart" class="chart-container"></div>
             </div>
-        </div>
-
-        <!-- 질문별 선택지 비율 차트: 한 줄에 2개씩 배치 -->
-        <div class="row row-custom">
             <div class="col-md-6">
-                <h3 class="text-center">1문항 선택지 비율</h3>
-                <div id="question_chart_0" class="chart-container"></div>
-            </div>
-            <div class="col-md-6">
-                <h3 class="text-center">2문항 선택지 비율</h3>
-                <div id="question_chart_1" class="chart-container"></div>
-            </div>
-        </div>
-
-        <div class="row row-custom">
-            <div class="col-md-6">
-                <h3 class="text-center">3문항 선택지 비율</h3>
-                <div id="question_chart_2" class="chart-container"></div>
-            </div>
-            <div class="col-md-6">
-                <h3 class="text-center">4문항 선택지 비율</h3>
-                <div id="question_chart_3" class="chart-container"></div>
+                <h3 class="text-center">연령별 사용자와 동일한 답변 선택자 비율</h3>
+                <div id="age_distribution_chart" class="chart-container"></div>
             </div>
         </div>
     </div>
